Renames the TypedTest class to TypedTestController and updates references
